refactor: move geocoder types to api folder
