+++ src/modules/geocoder/utils.ts
@@ -1,4 +1,4 @@
-import {Feature} from '@atb/sdk';
+import {Feature} from '@atb/api/bff/types';
 import {SearchLocation} from '@atb/modules/favorites';
 
 // IMPORTANT: Feature coordinate-array is [long, lat] :sadface:. Mapping to lat/long object for less bugs downstream.
